split types into files, add tests
Major types have been split into their own files to improve code
organization.

Start testing the library directly with a few tests, this
is far from being done but we have to start somewhere. Add a coverage
build step to build.zig using kcov.

libgit2 dependency has been reverted to the upstream allyourcodebases
version since it now supports Zig 0.14. The C version of the library
was also added to the dependencies in order to not depend on the system
`git2.h` header and instead import the one from the version defined
in `build.zig.zon`.